Fix problem on Raspberry Pi with unsigned char
On a Raspberry Pi (`gcc version 4.6.3 (Debian 4.6.3-14+rpi1)`) the following code causes a problem:
```
        char c;
        while ((c = getopt (argc, argv, "Z:z:o:h")) != -1) {
```
When getopt returned `-1`, it would be stored as `255` in `c` (treating `c` as an unsigned 8-bit value), but when `c` was compared with `-1`, it appears that `c` was treated as a regular int, and thus evaluated to `false`, triggering the default case and aborting. Switching the type of `c` to `int` fixes the problem.