Give users the ability to add IO read sources.
If you need to read a custom file format that wandio does not support
you can create a custom io_source_t for it, and add it to the list of
io_source_ts that wandio looks for when opening a file.

Custom sources are added with wandio_register_io_source, and need to
provide a pre_init function pointer that will recognize the custom
format. wandiocat has an example.